Added archival pre-GitHub files
These files are earlier versions of code that was eventually duplicated onto github. They appear to be snippets of test code that was later incorporated into "official" code versions at the creation of the repo, so they're included here purely for posterity, to ensure they are included at some point in the repo history, and may be deleted after the initial commit.